backend/django/requirements.txt
django>=5.0,<6.0
django-ninja>=1.0,<2.0
django-environ>=0.11.2
django-cors-headers>=4.3.1
ruff>=0.3.0
gunicorn>=21.2.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9  # Defaulting to psycopg2 for postgres, adjust if using other DBs
python-dotenv>=1.0.1
